// src/app/models/account-bank-request.ts
 export interface AccountBankRequest {
    adresseTitulaire?: string;
    codePostalTitulaire?: string;
    emailTitulaire?: string;
    nomTitulaire?: string;
    numeroTelephone?: string;
    paysTitulaire?: string;
    prenomTitulaire?: string;
    villeTitulaire?: string;
    typeCompte?: string;
    soldeInitial?: number;
    
 }